refactor(Map): Split out MapMarkers into separate file
